+++ src/util.h
@@ -5,7 +5,7 @@
 
 typedef std::basic_string<unsigned char> OctetString;
 
-std::string hexlify(const unsigned char *data, int length);
+std::string hexlify(const unsigned char *data, size_t length);
 std::string hexlify(const std::string &str);
 std::string hexlify(const OctetString &data);
